Fix the loading of the reference transform submodel.
Only the first "model =" line in a MODEL node has effect. To load multiple
models into a part, multiple MODEL nodes are required. This fixes the
problem of "Control from Here" not having the desired effect when used on
the docking ports. It does, however, break "Control From Here" for the
command module because ModuleCommand does not reset the part's reference
transform before setting the vessel's reference transform.