Create a module for setting the control reference.
This module makes it possible to use any transform of the part to which it
is attached as the reference transform ("Control From Here"). If the part
has a ModuleCommand module, then the command on the ModuleCommand module is
disabled in order to work around a stock bug where ModuleCommand does not
ensure the part's reference transform is correct before setting the
vessel's reference transform.